feat(k8s): dynamic account config (#3824)
* feat(k8s): dynamic account config

implement credential synchronization for kubernetes accounts. defines
`CredentialsInitializerSynchronizable` on the
`KubernetesV*ProviderConfig` classes. Each of these classes is
responsible for handling accounts for each provider version.

* feat(k8s): detemine if account has changed

* refactor(k8s): rename provider config classes

renames provider config classes to synchronizable since there is no
longer an `@Configuration` annotation.

* refactor(k8s): add file hash signature

adds a `kubeconfigFileHash` property which enables us to determine if
the file contents of the kubeconfig file changed even if the path hasn't
changed. this kind of change warrants an account refresh becuase the
credentials used by the account may have possibly changed.

* refactor(k8s): defer to computedOmitKinds

previously, we were modifying omitKinds when an account was initialized.
this caused calls to `equals` to fail because they didn't match
configured omitKinds. this change defers to `omitKindsComputed` so that
checking new and existing credentials is accurate according to how the
account it configured. this allows us to only initialize modified or new
accounts rather than every credential.