refactor(google): use Get instead of List to find a specific image (#3879)
I found this while looking for calls that needed pagination support. I
started to add pagination to the List call before I realized there was
really no point since we were applying a filter so that it would only
ever return a single image (or none).

I'm not sure why this didn't occur to me when I originally wrote the
code.