chore(dependencies): fix some accidental transitive dependencies (#863)
working on some cleanup in kork and testing it against various projects highlighed some sub-optimal
things in the dependency graph. Fixed those up by explicitly declaring some implementation dependencies
that were otherwise coming in as transitives of transitives of a kork module. Also cleaned up a few
spots where commons-lang 1 was used instead of commons-lang 3.

Also upgrades to latest gradle (5.5.1) to align with the version the spinnaker-gradle-plugin was
built against