fix(kayenta): plumb through siteLocal for kayenta email support (#3061)
This a temporary "hack" until https://github.com/spinnaker/kayenta/issues/600 is addressed.
This changes passes through the `siteLocal` stage context to the kayenta so that it can send
report emails. The `siteLocal` params (in `orca`) are currently used at Netflix only but they exist in
OSS `deck`. Medium term, we will remove this from deck and orca