feat(cloudformation): treat yaml as string instead of map (#3024)
To allow the template to contain intrinsic functions (e.g. !Ref)
this changes the processing of the template to consider the parameter
as a string instead of parsing it as a yaml object.

For backwards compatibility, if the value is a map then it will
be converted to a string before it's passed to the task.