refactor(gce): a lot of compute wrapper plumbing needed for the caching layer (#3970)
* refactor(gce): move the GoogleComputeOperations into helper classes

* refactor(gce): Add an executeGet method to get() compute requests

This returns an Optional<ResponseT> that is empty if the request got a
404.

* refactor(gce): Add a PaginatedComputeRequest for list() GCE calls

* refactor(gce): Add a way to run batches of paginated requests

This isn't necessarily the most efficient because we wait for ALL the
page(n) results to come in before sending off a batch request for
the page(n+1) results. I suppose you could request just the first pages
in a batch and request all subsequent pages in unbatched requests,
starting them immediately upon receipt of the previous page. But I think
this is probably good enough.

* refactor(gce): add more compute wrappers to be used by the caching layer

* refactor(gce): Use String#join instead of concatenation