fix(kubernetes): Add Json annotations (#3991)
These annotations were dropped in converting KubernetesKind;
this means that we're serializing a KubernetesKind as a full
object rather than just as a string. This slightly changes the
result of calls to /serverGroups and /loadBalancers as now the
'kind' field is an object not a string. This could cause issues
with consumers that expect a specific format; our integration tests
expect to see just a string there which is how I found this.