refactor(kubernetes): Fix some unsafe casts in credential creation (#4023)
* refactor(kubernetes): Split CredentialFactory to V1 and V2

A lot of the unecessary casts that we do when creating credentials
stem from the fact that we have a single CredentialFactory with
different methods for creating V1 and V2 credentials.

Instead, make a generic interface KubernetesCredentialFactory that
pulls out the common logic to default methods.  (These default methods
should probably both eventually live somewhere else, but I think
having them on the interface is a resonable solution for now.)

Then give both the V1 and V2 credentials an inner factory class that
implements that interface and allows us to create credentials in a
type-safe way. It also logically makes more sense, as the factory for
credentials lives with the credentials it's creating.

* refactor(kubernetes): Add some generic bounds to the account code

Add some parametrized types to eliminate a bunch of unsafe casts.

* refactor(kubernetes): Push getSpinnakerKindMap to V1/V2

The getSpinnakerKindMap function on the base
KubernetesNamedAccountCredentials class is needed to support
sending the kind map to the UI. It forks using an if block
depending on which implementation we're using, which means
we should push the logic down to the actual implementations.

One solution would have been to keep the kind map at the top
level and just implement an "add custom resources" in the
V2 provider class (that is a no-op in V1). I didn't do that
because:
(1) It's strange that there's logic in this class anyway, which
is just supposed to hold some metadata and delegate down to
the actual credentials.
(2) Ultimately, it probably makes sense for the V2 provider to
implement its own account-specific KubernetesSpinnakerKindMap
instead of building this map on the fly. Pushing this down to
the V1/V2 implementations allows us to do this later.

* fix(kubernetes): Add missing import statements