refactor(kubernetes): Move V1 code to clouddriver-kubernetes-v1 (#4028)
* refactor(kubernetes): Remove V1 dependence from V2 test

The V2-only KubernetesAccountResolverSpec depends on the
V1 credentials object KubernetesV1Credentials.  It's not actually
using any properties specific to the V1 credentials; this class
exists only to ensure that the resolver only returns V2 credentials.
Replace a mock for the V1 credentials with a mock for the interface,
which equivalently exercises the test case that we don't return
KubernetesCredentials that aren't KubernetesV2Credentials.

* refactor(kubernetes): Split KubernetesV1Configuration into its own file

In preparation for moving all V1-specific code to its own class, split out
the V1-only beans into KubernetesV1Configuration.

* refactor(kubernetes): Move V1 code to clouddriver-kubernetes-v1

This leaves only shared V1+V2 code in clouddriver-kubernetes, and
removes the dependency of the V2 provider on the V1 provider.

* refactor(kubernetes): Compile clouddriver-kubernetes with java

Now that the V1 code has been moved out of clouddriver-kubernetes,
only Java code is left. Rename the source directory from groovy
to java, and start compiling the module with the java compiler.