feat(rollback): avoid scaling down during rollback
Should partly address spinnaker/spinnaker#4895 by avoiding resizing the restore server group down.
This does not address concurrent actions issues, so it is probably still possible to have race conditions result in bad things.

So fundamentally this changes:
* computes the max of the current restore server group size and the rollback server group size
* only generates a resize stage if that computed size is bigger than the current size

This also opens the possibility to skip the snapshot/restoreMin tasks if we are not going to resize (which would pin the min of the restore server group).