chore(kubernetes): Bump fabric8 library to 4.6.0 (#4060)
* chore(kubernetes): Bump fabric8 library to 4.3.1

This commit also updates the code in response to a couple of
breaking changes in the upgrade. Notably, the cpu utilization target
of a replicaset is now a top-level field rather than being nested
in a metrics object. Likewise, the current utilization is a top-level
field of the status, rather than nested in a metrics object.

* chore(kubernetes): Bump fabric8 library to 4.6.0