feat(auth): Optionally skip authentication for image tagging (#3795)
* feat(auth): Optionally skip authentication for image tagging

Adds a new feature flag operations.security.allowUnauthenticatedImageTaggingInAccounts that takes a list of account names, and will skip the authentication check for tagging cloud images in AWS or GCE if the account name matches. The motivation is that if the bakery puts images in a protected account, any tagging stage will fail if the user or service account that triggered the pipeline does not have permissions to write to the account that contains the image, even if the same user just created the image. The feature defaults to disabled to not change the current behaviour.

* Move the functionality out of the controller and into the authorizer

Also make the feature more granular by taking a list of accounts instead of just a boolean.

* Fix wiring of descriptionAuthorizer