perf(kubernetes): Fix all agents handling on-demand requests (#4110)
* test(kubernetes): Add tests to KubernetesCoreCachingAgent

The next commit will fix a bug in the Kubernetes caching agents;
there are currently no tests validating these agents. Add some tests
on the KubernetesCoreCachingAgent.

* perf(kubernetes): Fix all agents handling on-demand requests

Currently sending an on-demand cache update for a resource that
is not namespaced will cause all caching agents to try to handle
the update. This can cause a very high load on redis if there are
a lot of agents (as might be expected in a large installation of
Spinnaker).

Designate one agent to handle requests that are not namespaced
instead of having them handled by all agents.

* test(kubernetes): Improve KubernetesCoreCachingAgent tests

Instead of mocking the ProviderCache, use an in-memory cache, and
validate that the cache has the expected results after on-demand
requests have been processed.