chore(dependencies): fiatVersion 1.7.0 (#529)
fixes a test that broke as a result of coupling the API spec to the
fiat Permissions object. This just fixes the resulting change, but
a proper refactor would be to maybe undo that in favor of something
that translated to and from Map<String, List<String>> to avoid
future breakages if/when additional authorization types are added.