fix(pipeline): update invokePipelineConfig for proper return (#885)
Related to spinnaker/spin#204

We are using the spin-cli package as an API wrapper and ran in to a problem with the return value of InvokePipelineConfigUsingPOST1 not including the pipeline execution ID. This turned out to be because the response was being marshaled in to an HttpEntity which didn't have the ref JSON. Swagger generated this based on the fact that the return for invokePipelineConfig was HttpEntity. So this updates that method to return a Map instead which will cause Swagger to generate a new schema to marshal the return value in spin-cli to an empty struct.